refactor(de): Pull put datetime code
