+++ crates/toml_edit/src/de/value.rs
@@ -1,5 +1,6 @@
 use serde::de::IntoDeserializer;
 
+use crate::de::DatetimeDeserializer;
 use crate::de::Error;
 
 impl<'de> serde::Deserializer<'de> for crate::Value {
@@ -14,10 +15,9 @@ impl<'de> serde::Deserializer<'de> for crate::Value {
             crate::Value::Integer(v) => visitor.visit_i64(v.into_value()),
             crate::Value::Float(v) => visitor.visit_f64(v.into_value()),
             crate::Value::Boolean(v) => visitor.visit_bool(v.into_value()),
-            crate::Value::Datetime(v) => visitor.visit_map(DatetimeDeserializer {
-                date: v.into_value(),
-                visited: false,
-            }),
+            crate::Value::Datetime(v) => {
+                visitor.visit_map(DatetimeDeserializer::new(v.into_value()))
+            }
             crate::Value::Array(v) => v.into_deserializer().deserialize_any(visitor),
             crate::Value::InlineTable(v) => v.into_deserializer().deserialize_any(visitor),
         }
@@ -34,10 +34,7 @@ impl<'de> serde::Deserializer<'de> for crate::Value {
     {
         if name == toml_datetime::__unstable::NAME && fields == [toml_datetime::__unstable::FIELD] {
             if let crate::Value::Datetime(d) = self {
-                return visitor.visit_map(DatetimeDeserializer {
-                    date: d.into_value(),
-                    visited: false,
-                });
+                return visitor.visit_map(DatetimeDeserializer::new(d.into_value()));
             }
         }
 
@@ -106,51 +103,5 @@ impl<'de> serde::de::IntoDeserializer<'de, crate::de::Error> for crate::Value {
 
     fn into_deserializer(self) -> Self::Deserializer {
         self
-    }
-}
-
-struct DatetimeDeserializer {
-    visited: bool,
-    date: crate::Datetime,
-}
-
-impl<'de> serde::de::MapAccess<'de> for DatetimeDeserializer {
-    type Error = Error;
-
-    fn next_key_seed<K>(&mut self, seed: K) -> Result<Option<K::Value>, Error>
-    where
-        K: serde::de::DeserializeSeed<'de>,
-    {
-        if self.visited {
-            return Ok(None);
-        }
-        self.visited = true;
-        seed.deserialize(DatetimeFieldDeserializer).map(Some)
-    }
-
-    fn next_value_seed<V>(&mut self, seed: V) -> Result<V::Value, Error>
-    where
-        V: serde::de::DeserializeSeed<'de>,
-    {
-        seed.deserialize(self.date.to_string().into_deserializer())
-    }
-}
-
-struct DatetimeFieldDeserializer;
-
-impl<'de> serde::de::Deserializer<'de> for DatetimeFieldDeserializer {
-    type Error = Error;
-
-    fn deserialize_any<V>(self, visitor: V) -> Result<V::Value, Error>
-    where
-        V: serde::de::Visitor<'de>,
-    {
-        visitor.visit_borrowed_str(toml_datetime::__unstable::FIELD)
-    }
-
-    serde::forward_to_deserialize_any! {
-        bool u8 u16 u32 u64 i8 i16 i32 i64 f32 f64 char str string seq
-        bytes byte_buf map struct option unit newtype_struct
-        ignored_any unit_struct tuple_struct tuple enum identifier
     }
 }
